requirements.txt
instaloader
pymongo>=4.6.0
requests>=2.31.0
python-dotenv>=1.0.0
